% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo_year_df.r
\docType{data}
\name{cargo_year_df}
\alias{cargo_year_df}
\title{cargo_year_df}
\format{
A data frame comprising 880597 rows and 4 columns. 
\describe{
   \item{year}{year of route density data}
   \item{x}{longitude}
   \item{y}{latitude}
   \item{value}{Cargo shipping route density/km2}
   }
}
\usage{
cargo_year_df
}
\description{
Cargo shipping route density through the Black Sea to August 2021 and to December 2022. Data originally from EMODNET.
}
\keyword{datasets}
