% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ukr_FAOSTAT.r
\docType{data}
\name{Ukr_FAOSTAT}
\alias{Ukr_FAOSTAT}
\title{Ukr_FAOSTAT}
\format{
A data frame comprising 198 rows and 14 columns. 
\describe{
   \item{Domain.Code}{Domain.Code}
   \item{Domain}{Domain}
   \item{Area.Code..FAO.}{Area.Code..FAO.}
   \item{Area}{Area}
   \item{Element.Code}{Element.Code}
   \item{Element}{Element}
   \item{Item.Code..FAO.}{Item.Code..FAO.}
   \item{Item}{Item}
   \item{Year.Code}{Year.Code}
   \item{Year}{Year}
   \item{Unit}{Unit}  
   \item{Value}{Value}          
   \item{Flag}{Flag}
   \item{Flag.Description}{Flag.Description}    
   }
}
\usage{
Ukr_FAOSTAT
}
\description{
Crop production for 2010-2020 for Ukraine originally taken from FAOSTAT.
}
\keyword{datasets}
